bitwrap/requirements.txt
Flask>=1.0.2
Flask-SocketIO>=3.0.0
Flask-Cors>=3.0.4
Flask-RESTful>=0.3.6
GitHub-Flask>=3.2.0
psycopg2-binary>=2.7.4
